Updated settings files.
All of the settings available under the "Settings" tab in the WordPress
administration section have been worked through and mirrored in
installation script. This needs to be completed by emphasizing the more
important elements of the site (such as site title and description), and
by creating relevant installation functions.  All of the folders, files
and data files should be in place.